generate code automatic
